backend/requirements.txt
langchain
langchain-openai~=0.3.14
pinecone~=6.0.2
langchain-pinecone
langchain_community
fastapi~=0.115.12
pypdf
langchain-community~=0.3.23
python-dotenv~=1.1.0
langchain-text-splitters~=0.3.8
pinecone-datasets